payable-features contract extracted from basic-features
