+++ contracts/feature-tests/payable-features/.gitignore
@@ -1,0 +1,7 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+*/target/
+
+# The erdpy output
+output
