+++ contracts/feature-tests/payable-features/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "payable-features"
+version = "0.0.0"
+authors = ["Andrei Marinica <andrei.marinica@elrond.com>"]
+edition = "2018"
+publish = false
+
+[lib]
+path = "src/lib.rs"
+
+[features]
+wasm-output-mode = ["elrond-wasm-node"]
+
+[dependencies.elrond-wasm]
+version = "0.13.0"
+path = "../../../elrond-wasm"
+
+[dependencies.elrond-wasm-derive]
+version = "0.13.0"
+path = "../../../elrond-wasm-derive"
+
+[dependencies.elrond-wasm-node]
+version = "0.13.0"
+path = "../../../elrond-wasm-node"
+optional = true
+
+[dev-dependencies.elrond-wasm-debug]
+version = "0.13.0"
+path = "../../../elrond-wasm-debug"
