+++ contracts/feature-tests/payable-features/abi/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "payable-features-abi"
+version = "0.0.0"
+authors = ["Andrei Marinica <andrei.marinica@elrond.com>"]
+edition = "2018"
+publish = false
+
+[dependencies.payable-features]
+path = ".."
+
+[dependencies.elrond-wasm]
+version = "0.13.0"
+path = "../../../../elrond-wasm"
+
+[dependencies.elrond-wasm-debug]
+version = "0.13.0"
+path = "../../../../elrond-wasm-debug"
