+++ contracts/feature-tests/payable-features/elrond.json
@@ -1,0 +1,3 @@
+{
+    "language": "rust"
+}
