+++ contracts/feature-tests/payable-features/mandos/payable_any_2.scen.json
@@ -9,14 +9,14 @@
                     "nonce": "0",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "code": "file:../output/payable-features.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
                     "balance": "1,000,000,000,000",
                     "storage": {},
                     "esdt": {
-                        "str:BASIC-FEATURES-TOKEN": "1,000,000,000,000",
+                        "str:PAYABLE-FEATURES-TOKEN": "1,000,000,000,000",
                         "str:OTHER-TOKEN": "1,000,000,000,000"
                     },
                     "code": ""
@@ -25,12 +25,12 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_any_1.1",
+            "txId": "payable_any_2.1",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "payable_any_1",
+                "function": "payable_any_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -45,12 +45,12 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_any_1.2",
+            "txId": "payable_any_2.2",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "5",
-                "function": "payable_any_1",
+                "function": "payable_any_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -65,20 +65,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_any_1.3",
+            "txId": "payable_any_2.3",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "100",
-                "function": "payable_any_1",
+                "function": "payable_any_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "100", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "100", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -87,20 +87,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_any_1.4",
+            "txId": "payable_any_2.4",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "0",
-                "function": "payable_any_1",
+                "function": "payable_any_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "0", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "0", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
