+++ contracts/feature-tests/payable-features/mandos/payable_any_3.scen.json
@@ -9,14 +9,14 @@
                     "nonce": "0",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "code": "file:../output/payable-features.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
                     "balance": "1,000,000,000,000",
                     "storage": {},
                     "esdt": {
-                        "str:BASIC-FEATURES-TOKEN": "1,000,000,000,000",
+                        "str:PAYABLE-FEATURES-TOKEN": "1,000,000,000,000",
                         "str:OTHER-TOKEN": "1,000,000,000,000"
                     },
                     "code": ""
@@ -25,20 +25,19 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_4.1",
+            "txId": "payable_any_3.1",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "payable_token_4",
+                "function": "payable_any_3",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
+                "out": [ "0", "str:EGLD" ],
+                "status": "",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
@@ -46,41 +45,18 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_4.2",
+            "txId": "payable_any_3.2",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "5",
-                "function": "payable_token_4",
+                "function": "payable_any_3",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
-                "logs": [],
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "payable_token_4.3",
-            "tx": {
-                "from": "address:an_account",
-                "to": "address:features_contract",
-                "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "100",
-                "function": "payable_token_4",
-                "arguments": [],
-                "gasLimit": "50,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "out": [ "100", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "5", "str:EGLD" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -89,20 +65,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_4.4",
+            "txId": "payable_any_3.3",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "0",
-                "function": "payable_token_4",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
+                "esdtValue": "100",
+                "function": "payable_any_3",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "0", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "100", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -111,45 +87,21 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_4.5",
+            "txId": "payable_any_3.4",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:OTHER-TOKEN",
-                "esdtValue": "100",
-                "function": "payable_token_4",
-                "arguments": [],
-                "gasLimit": "50,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
-                "logs": [],
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "payable_token_4.6",
-            "tx": {
-                "from": "address:an_account",
-                "to": "address:features_contract",
-                "value": "0",
-                "esdtTokenName": "str:OTHER-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "0",
-                "function": "payable_token_4",
+                "function": "payable_any_3",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
+                "out": [ "0", "str:PAYABLE-FEATURES-TOKEN" ],
+                "status": "",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
