+++ contracts/feature-tests/payable-features/mandos/payable_egld_2.scen.json
@@ -9,14 +9,14 @@
                     "nonce": "0",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "code": "file:../output/payable-features.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
                     "balance": "1,000,000,000,000",
                     "storage": {},
                     "esdt": {
-                        "str:BASIC-FEATURES-TOKEN": "1,000,000,000,000",
+                        "str:PAYABLE-FEATURES-TOKEN": "1,000,000,000,000",
                         "str:OTHER-TOKEN": "1,000,000,000,000"
                     },
                     "code": ""
@@ -25,20 +25,19 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_1.1",
+            "txId": "payable_egld_2.1",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "payable_token_1",
+                "function": "payable_egld_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
+                "out": [ "0", "str:EGLD" ],
+                "status": "",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
@@ -46,41 +45,18 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_1.2",
+            "txId": "payable_egld_2.2",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "5",
-                "function": "payable_token_1",
+                "function": "payable_egld_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:bad call value token provided",
-                "logs": [],
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "payable_token_1.3",
-            "tx": {
-                "from": "address:an_account",
-                "to": "address:features_contract",
-                "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "100",
-                "function": "payable_token_1",
-                "arguments": [],
-                "gasLimit": "50,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "out": [ "100", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "5", "str:EGLD" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -89,36 +65,14 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_1.4",
+            "txId": "payable_egld_2.3",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "0",
-                "function": "payable_token_1",
-                "arguments": [],
-                "gasLimit": "50,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "out": [ "0", "str:BASIC-FEATURES-TOKEN" ],
-                "status": "",
-                "logs": [],
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "payable_token_1.5",
-            "tx": {
-                "from": "address:an_account",
-                "to": "address:features_contract",
-                "value": "0",
-                "esdtTokenName": "str:OTHER-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "100",
-                "function": "payable_token_1",
+                "function": "payable_egld_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -126,7 +80,7 @@
             "expect": {
                 "out": [],
                 "status": "4",
-                "message": "str:bad call value token provided",
+                "message": "str:function does not accept ESDT payment",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
@@ -134,14 +88,14 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_1.6",
+            "txId": "payable_egld_2.4",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:OTHER-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "0",
-                "function": "payable_token_1",
+                "function": "payable_egld_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -149,7 +103,7 @@
             "expect": {
                 "out": [],
                 "status": "4",
-                "message": "str:bad call value token provided",
+                "message": "str:function does not accept ESDT payment",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
