+++ contracts/feature-tests/payable-features/mandos/payable_token_1.scen.json
@@ -9,14 +9,14 @@
                     "nonce": "0",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "code": "file:../output/payable-features.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
                     "balance": "1,000,000,000,000",
                     "storage": {},
                     "esdt": {
-                        "str:BASIC-FEATURES-TOKEN": "1,000,000,000,000",
+                        "str:PAYABLE-FEATURES-TOKEN": "1,000,000,000,000",
                         "str:OTHER-TOKEN": "1,000,000,000,000"
                     },
                     "code": ""
@@ -25,19 +25,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_egld_3.1",
+            "txId": "payable_token_1.1",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "payable_egld_3",
+                "function": "payable_token_1",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "0", "str:EGLD" ],
-                "status": "",
+                "out": [],
+                "status": "4",
+                "message": "str:bad call value token provided",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
@@ -45,18 +46,41 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_egld_3.2",
+            "txId": "payable_token_1.2",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "5",
-                "function": "payable_egld_3",
+                "function": "payable_token_1",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "5", "str:EGLD" ],
+                "out": [],
+                "status": "4",
+                "message": "str:bad call value token provided",
+                "logs": [],
+                "gas": "*",
+                "refund": "*"
+            }
+        },
+        {
+            "step": "scCall",
+            "txId": "payable_token_1.3",
+            "tx": {
+                "from": "address:an_account",
+                "to": "address:features_contract",
+                "value": "0",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
+                "esdtValue": "100",
+                "function": "payable_token_1",
+                "arguments": [],
+                "gasLimit": "50,000,000",
+                "gasPrice": "0"
+            },
+            "expect": {
+                "out": [ "100", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -65,22 +89,21 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_egld_3.3",
+            "txId": "payable_token_1.4",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "100",
-                "function": "payable_egld_3",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
+                "esdtValue": "0",
+                "function": "payable_token_1",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [],
-                "status": "4",
-                "message": "str:function does not accept ESDT payment",
+                "out": [ "0", "str:PAYABLE-FEATURES-TOKEN" ],
+                "status": "",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
@@ -88,14 +111,14 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_egld_3.4",
+            "txId": "payable_token_1.5",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
-                "esdtValue": "0",
-                "function": "payable_egld_3",
+                "esdtTokenName": "str:OTHER-TOKEN",
+                "esdtValue": "100",
+                "function": "payable_token_1",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -103,7 +126,30 @@
             "expect": {
                 "out": [],
                 "status": "4",
-                "message": "str:function does not accept ESDT payment",
+                "message": "str:bad call value token provided",
+                "logs": [],
+                "gas": "*",
+                "refund": "*"
+            }
+        },
+        {
+            "step": "scCall",
+            "txId": "payable_token_1.6",
+            "tx": {
+                "from": "address:an_account",
+                "to": "address:features_contract",
+                "value": "0",
+                "esdtTokenName": "str:OTHER-TOKEN",
+                "esdtValue": "0",
+                "function": "payable_token_1",
+                "arguments": [],
+                "gasLimit": "50,000,000",
+                "gasPrice": "0"
+            },
+            "expect": {
+                "out": [],
+                "status": "4",
+                "message": "str:bad call value token provided",
                 "logs": [],
                 "gas": "*",
                 "refund": "*"
