+++ contracts/feature-tests/payable-features/mandos/payable_token_2.scen.json
@@ -9,14 +9,14 @@
                     "nonce": "0",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "code": "file:../output/payable-features.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
                     "balance": "1,000,000,000,000",
                     "storage": {},
                     "esdt": {
-                        "str:BASIC-FEATURES-TOKEN": "1,000,000,000,000",
+                        "str:PAYABLE-FEATURES-TOKEN": "1,000,000,000,000",
                         "str:OTHER-TOKEN": "1,000,000,000,000"
                     },
                     "code": ""
@@ -25,12 +25,12 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.1",
+            "txId": "payable_token_2.1",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -46,12 +46,12 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.2",
+            "txId": "payable_token_2.2",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "5",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -67,20 +67,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.3",
+            "txId": "payable_token_2.3",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "100",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "100", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "100", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -89,20 +89,20 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.4",
+            "txId": "payable_token_2.4",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "esdtTokenName": "str:BASIC-FEATURES-TOKEN",
+                "esdtTokenName": "str:PAYABLE-FEATURES-TOKEN",
                 "esdtValue": "0",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
             },
             "expect": {
-                "out": [ "0", "str:BASIC-FEATURES-TOKEN" ],
+                "out": [ "0", "str:PAYABLE-FEATURES-TOKEN" ],
                 "status": "",
                 "logs": [],
                 "gas": "*",
@@ -111,14 +111,14 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.5",
+            "txId": "payable_token_2.5",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
                 "esdtTokenName": "str:OTHER-TOKEN",
                 "esdtValue": "100",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
@@ -134,14 +134,14 @@
         },
         {
             "step": "scCall",
-            "txId": "payable_token_3.6",
+            "txId": "payable_token_2.6",
             "tx": {
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
                 "esdtTokenName": "str:OTHER-TOKEN",
                 "esdtValue": "0",
-                "function": "payable_token_3",
+                "function": "payable_token_2",
                 "arguments": [],
                 "gasLimit": "50,000,000",
                 "gasPrice": "0"
