+++ contracts/feature-tests/payable-features/wasm/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "payable-features-wasm"
+version = "0.0.0"
+authors = ["Andrei Marinica <andrei.marinica@elrond.com>"]
+edition = "2018"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[profile.release]
+codegen-units = 1
+opt-level = "z"
+lto = true
+debug = false
+panic = "abort"
+
+[dependencies.payable-features]
+path = ".."
+features = ["wasm-output-mode"]
+default-features = false
+
+[dependencies.elrond-wasm-output]
+version = "0.13.0"
+path = "../../../../elrond-wasm-output"
+features = ["wasm-output-mode"]
+
+[workspace]
+members = ["."]
