+++ contracts/feature-tests/payable-features/wasm/src/lib.rs
@@ -1,0 +1,5 @@
+
+#![no_std]
+
+pub use payable_features::*;
+pub use elrond_wasm_output::*;
